+++ modelcraft/modelcraft.py
@@ -1,6 +1,5 @@
 import json
 import os
-import re
 import sys
 import time
 import gemmi
@@ -23,8 +22,11 @@ from .structure import ModelStats, remove_residues, write_mmcif
 class ModelCraft(Pipeline):
     def __init__(self, args):
         self.args = parse(args)
-        print(f"# ModelCraft {__version__}")
-        super().__init__(keep_jobs=self.args.keep_files, keep_logs=self.args.keep_logs)
+        super().__init__(
+            directory=self.args.directory,
+            keep_jobs=self.args.keep_files,
+            keep_logs=self.args.keep_logs,
+        )
         self.cycle = 0
         self.current_structure: gemmi.Structure = self.args.model
         self.current_phases: DataItem = self.args.phases
@@ -42,28 +44,15 @@ class ModelCraft(Pipeline):
 
     @property
     def resolution(self):
-        return self.args.fsigf.resolution_high()
+        return self.args.fmean.resolution_high()
 
     def run(self):
         self.start_time = time.time()
         args = self.args
-        os.makedirs(args.directory, exist_ok=True)
-        os.chdir(args.directory)
-        _check_for_files_that_could_be_overwritten()
-        if self.args.observations.types == "FQ":
-            self.args.fsigf = self.args.observations
-        else:
-            print("\n## Converting input observations to mean amplitudes\n")
-            result = CTruncate(observations=self.args.observations).run(self)
-            self.args.fsigf = result.fmean
-        if args.mode == "xray" and args.model is not None:
-            print("\n## Refining Input Model\n")
-            self.update_model_cell()
-            self.sheetbend()
-            args.model = self.current_structure
-            if args.phases is not None:
-                self.current_phases = args.phases
-            self.print_refmac_result(self.last_refmac)
+        print(f"# ModelCraft {__version__}")
+        os.makedirs(args.directory, exist_ok=False)
+        self._convert_observations()
+        self._refine_input_model()
         for self.cycle in range(1, args.cycles + 1):
             print("\n## Cycle %d\n" % self.cycle)
             self.run_cycle()
@@ -87,6 +76,27 @@ class ModelCraft(Pipeline):
         print("\n## Best Model:")
         self.print_refmac_result(self.output_refmac)
         self.terminate(reason="Normal")
+
+    def _convert_observations(self):
+        if self.args.fmean is None:
+            print("\n## Converting input observations to mean amplitudes\n")
+            observations = self.args.ianom or self.args.imean or self.args.fanom
+            ctruncate = CTruncate(observations=observations).run(self)
+            self.args.fmean = ctruncate.fmean
+            if self.args.fanom is None and ctruncate.fanom is not None:
+                self.args.fanom = ctruncate.fanom
+            if self.args.imean is None and ctruncate.imean is not None:
+                self.args.imean = ctruncate.imean
+
+    def _refine_input_model(self):
+        if self.args.mode == "xray" and self.args.model is not None:
+            print("\n## Refining Input Model\n")
+            self.update_model_cell()
+            self.sheetbend()
+            self.args.model = self.current_structure
+            if self.args.phases is not None:
+                self.current_phases = self.args.phases
+            self.print_refmac_result(self.last_refmac)
 
     def run_cycle(self):
         if self.args.mode == "em":
@@ -119,7 +129,7 @@ class ModelCraft(Pipeline):
     def sheetbend(self):
         print("Sheetbend")
         result = Sheetbend(
-            fsigf=self.args.fsigf,
+            fsigf=self.args.fmean,
             freer=self.args.freer,
             structure=self.current_structure,
         ).run(self)
@@ -131,7 +141,7 @@ class ModelCraft(Pipeline):
         print("Buccaneer")
         result = Buccaneer(
             contents=self.args.contents,
-            fsigf=self.args.fsigf,
+            fsigf=self.args.fmean,
             phases=self.current_phases,
             fphi=self.current_fphi_best if self.args.mode == "xray" else None,
             freer=self.args.freer if self.args.mode == "xray" else None,
@@ -153,7 +163,7 @@ class ModelCraft(Pipeline):
         print("Nautilus")
         result = Nautilus(
             contents=self.args.contents,
-            fsigf=self.args.fsigf,
+            fsigf=self.args.fmean,
             phases=self.current_phases,
             fphi=self.current_fphi_best if self.args.mode == "xray" else None,
             freer=self.args.freer if self.args.mode == "xray" else None,
@@ -168,7 +178,7 @@ class ModelCraft(Pipeline):
             )
             result = RefmacXray(
                 structure=structure,
-                fsigf=self.args.fsigf,
+                fsigf=self.args.fmean,
                 freer=self.args.freer,
                 cycles=cycles,
                 phases=self.args.phases if use_phases else None,
@@ -209,7 +219,7 @@ class ModelCraft(Pipeline):
         print("Parrot")
         result = Parrot(
             contents=self.args.contents,
-            fsigf=self.args.fsigf,
+            fsigf=self.args.fmean,
             freer=self.args.freer,
             phases=self.current_phases,
             fphi=self.current_fphi_best,
@@ -264,11 +274,11 @@ class ModelCraft(Pipeline):
         if self._is_better(self.last_refmac, self.output_refmac):
             self.cycles_without_improvement = 0
             self.output_refmac = self.last_refmac
-            write_mmcif("modelcraft.cif", self.last_refmac.structure)
+            write_mmcif(self.path("modelcraft.cif"), self.last_refmac.structure)
             write_mtz(
-                "modelcraft.mtz",
+                self.path("modelcraft.mtz"),
                 [
-                    self.args.fsigf,
+                    self.args.fmean,
                     self.args.freer,
                     self.last_refmac.abcd,
                     self.last_refmac.fphi_best,
@@ -283,7 +293,7 @@ class ModelCraft(Pipeline):
 
     def write_report(self):
         self.seconds["total"] = time.time() - self.start_time
-        with open("modelcraft.json", "w") as report_file:
+        with open(self.path("modelcraft.json"), "w") as report_file:
             json.dump(self.report, report_file, indent=4)
 
     def print_refmac_result(self, result: RefmacResult):
@@ -298,7 +308,7 @@ class ModelCraft(Pipeline):
 
     def update_model_cell(self):
         structure = self.args.model
-        mtz = self.args.fsigf
+        mtz = self.args.fmean
         structure_spacegroup = gemmi.find_spacegroup_by_name(
             structure.spacegroup_hm,
             alpha=structure.cell.alpha,
@@ -317,21 +327,3 @@ class ModelCraft(Pipeline):
             self.terminate("Model cell is incompatible")
         remove_scale(structure=structure)
         update_cell(structure=structure, new_cell=mtz.cell)
-
-
-def _check_for_files_that_could_be_overwritten():
-    patterns = [
-        r"modelcraft\.cif",
-        r"modelcraft\.json",
-        r"modelcraft\.mtz",
-        r"job_[A-Za-z0-9]+_[A-Za-z0-9]{20}",
-        r"job_\d+_[A-Za-z0-9]+",
-    ]
-    paths = os.listdir(".")
-    paths = [p for p in paths if any(re.fullmatch(pattern, p) for pattern in patterns)]
-    if paths:
-        print("\nThe following files may be from a previous run:\n")
-        for path in paths:
-            print("-", path)
-        print("\nPlease run in a different directory or remove these files.")
-        sys.exit()