+++ modelcraft/tests/integration/test_comit.py
@@ -1,10 +1,15 @@
+import gemmi
 from modelcraft.jobs.comit import Comit
-from . import insulin_fsigf, insulin_refmac
+from modelcraft.reflections import DataItem
+from . import ccp4_path
 
 
-def test_insulin():
-    fsigf = insulin_fsigf()
-    refmac = insulin_refmac()
-    comit = Comit(fsigf, refmac.fphi_best).run()
-    assert comit.abcd.nreflections == fsigf.nreflections
-    assert comit.fphi.nreflections == fsigf.nreflections
+def test_comit_after_comit():
+    mtz_path = ccp4_path("examples", "data", "gere.mtz")
+    mtz = gemmi.read_mtz_file(mtz_path)
+    fsigf = DataItem(mtz, "FPHASED,SIGFPHASED")
+    fphi = DataItem(mtz, "FB,PHIB")
+    comit1 = Comit(fsigf, fphi).run()
+    comit2 = Comit(fsigf, comit1.fphi).run()
+    assert comit2.abcd.nreflections == fsigf.nreflections
+    assert comit2.fphi.nreflections == fsigf.nreflections